requirements.txt
asgiref==3.7.2
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.10
Pillow==9.5.0
pytils==0.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
